Disallows mappings outside storage.

+++ test/libsolidity/syntaxTests/types/mapping/memory_structs_with_mappings.sol
@@ -1,16 +1,10 @@
 contract Test {
-    struct S {
-        uint8 a;
-        mapping(uint256 => uint256) b;
-        uint8 c;
-    }
+    struct S { uint8 a; mapping(uint => uint) b; uint8 c; }
     S s;
-
-    function f() public returns (uint256) {
+    function f() public returns (uint) {
         S memory x;
         if (x.a != 0 || x.c != 0) return 1;
-        x.a = 4;
-        x.c = 5;
+        x.a = 4; x.c = 5;
         s = x;
         if (s.a != 4 || s.c != 5) return 2;
         x = S(2, 3);
@@ -19,6 +13,5 @@ contract Test {
         if (s.a != 4 || s.c != 5) return 4;
     }
 }
-
 // ----
-// f() -> 0
+// TypeError: (134-144): Type struct Test.S memory is only valid in storage.
